See if the wrong type of column was used in a migration

+++ database/migrations/2021_03_16_130344_add_programme_id_to_users_table.php
@@ -14,7 +14,7 @@ class AddProgrammeIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('programme_id')->nullable();
+            $table->unsignedInteger('programme_id')->nullable();
             $table->foreign('programme_id')->references('id')->on('programmes');
         });
     }
